move tests for areRoleRightsValid and IsRoleRights to _/decorator_ folder
